componentized theme toggle
